fixed SetColor function & added stamina feedback to the particle

+++ Ko-Fi Engine/Source/C_Particle.h
@@ -35,7 +35,7 @@ public:
 	void ResumeParticleSpawn();
 	void ResetTimers();
 	void SetLoop(bool v);
-	void SetColor(float r, float g, float b, float a);
+	void SetColor(double r, double g, double b, double a);
 	void DeleteModule(Emitter* e,ParticleModuleType t);
 
 private:
